[api] Add missing fileUniqueId and migrate id from Int to Long
The fileUniqueId field was added on many models in the API.
According to Telegram's latest announcement (see below), the user identifier might not fit in a Int anymore, we migrated our schema to use a Long instead.

warning WARNING! warning
After one of the upcoming Bot API updates, some user identifiers will become bigger than 2^31 - 1 and it will be no longer possible to store them in a signed 32-bit integer type. User identifiers will have up to 52 significant bits, so a 64-bit integer or double-precision float type would still be safe for storing them. Please make sure that your code can correctly handle such user identifiers.

Fixes #91

+++ examples/src/AuthenticationBot.scala
@@ -10,7 +10,7 @@ import scala.concurrent.Future
   * Extension to add a simple authentication filter.
   */
 trait SillyAuthentication {
-  val allowed = scala.collection.mutable.Set[Int]()
+  val allowed = scala.collection.mutable.Set[Long]()
 
   def login(user: User) = atomic {
     allowed += user.id
